requirements.txt
pytesseract~=0.3.10
opencv-python~=4.9.0.80
flask~=3.0.2
pillow~=10.2.0